Refactor rewards calculation module: update imports and restructure code for clarity
