+++ rs/dre-canisters/node-provider-rewards/canister/src/metrics.rs
@@ -6,8 +6,9 @@ use ic_cdk::api::call::CallResult;
 use ic_management_canister_types_private::{NodeMetricsHistoryArgs, NodeMetricsHistoryResponse};
 use ic_stable_structures::StableBTreeMap;
 use itertools::Itertools;
+use rewards_calculation::input_builder::NodeMetricsDaily;
 use rewards_calculation::metrics::NodeMetricsDaily;
-use rewards_calculation::reward_period::TimestampNanos;
+use rewards_calculation::types::TimestampNanos;
 use std::cell::RefCell;
 use std::collections::{BTreeMap, HashMap};
