+++ rs/dre-canisters/node-provider-rewards/canister/src/metrics_types.rs
@@ -2,7 +2,7 @@ use candid::{CandidType, Decode, Encode, Principal};
 use ic_base_types::{NodeId, PrincipalId, SubnetId};
 use ic_stable_structures::storable::Bound;
 use ic_stable_structures::Storable;
-use rewards_calculation::reward_period::TimestampNanos;
+use rewards_calculation::types::TimestampNanos;
 use serde::Deserialize;
 use std::borrow::Cow;
