+++ rs/dre-canisters/node-provider-rewards/rewards-calculation/src/lib.rs
@@ -1,26 +1,12 @@
-use crate::calculation_results::NodeProviderCalculationResults;
-use crate::execution_context::{nodes_ids, ExecutionContext};
-use crate::metrics::{nodes_failure_rates_in_period, subnets_failure_rates, NodeMetricsDaily, SubnetDailyFailureRate};
-use crate::reward_period::{RewardPeriod, TimestampNanos};
-use crate::types::{rewardable_nodes_by_provider, RewardableNode};
-use ic_base_types::{NodeId, PrincipalId, SubnetId};
-use ic_protobuf::registry::node_rewards::v2::NodeRewardsTable;
-use std::cmp::PartialEq;
-use std::collections::{BTreeMap, HashSet};
-use std::error::Error;
-use std::fmt;
+use crate::calculation_results::RewardsCalculatorResults;
+use crate::input_builder::{RewardCalculationError, RewardableNode, RewardsCalculatorInput};
+use crate::rewards_calculator::{Initialized, RewardsCalculator, RewardsTotalComputed};
+use std::marker::PhantomData;
 
 pub mod calculation_results;
-mod execution_context;
-pub mod metrics;
-pub mod reward_period;
+pub mod input_builder;
+mod rewards_calculator;
 pub mod types;
-
-#[derive(Default)]
-pub struct RewardsCalculationResult {
-    pub subnets_failure_rates: BTreeMap<SubnetId, Vec<SubnetDailyFailureRate>>,
-    pub results_per_provider: BTreeMap<PrincipalId, NodeProviderCalculationResults>,
-}
 
 /// Computes rewards for node providers based on their nodes' performance during the specified `reward_period`.
 ///
@@ -30,112 +16,24 @@ pub struct RewardsCalculationResult {
 /// * metrics_by_node - Daily node metrics for nodes in `reward_period`. Only nodes in `providers_rewardable_nodes` keys are considered.
 /// * rewardable_nodes: Nodes eligible for rewards, as recorded in the registry versions spanning the `reward_period` provided.
 pub fn calculate_rewards(
-    reward_period: &RewardPeriod,
-    rewards_table: &NodeRewardsTable,
-    daily_metrics_by_node: &BTreeMap<NodeId, Vec<NodeMetricsDaily>>,
-    rewardable_nodes: &[RewardableNode],
-) -> Result<RewardsCalculationResult, RewardCalculationError> {
-    let mut rewards_calculation_result = RewardsCalculationResult::default();
-    let all_nodes = nodes_ids(rewardable_nodes);
-
-    validate_input(reward_period, daily_metrics_by_node, &all_nodes)?;
-
-    let ctx = ExecutionContext::new(
-        nodes_failure_rates_in_period(&all_nodes, reward_period, daily_metrics_by_node),
-        subnets_failure_rates(daily_metrics_by_node),
-        rewards_table.clone(),
-    );
-
-    for (provider_id, provider_nodes) in rewardable_nodes_by_provider(rewardable_nodes) {
-        let node_provider_results: NodeProviderCalculationResults = ctx.calculate_rewards(provider_nodes);
-
-        rewards_calculation_result.results_per_provider.insert(provider_id, node_provider_results);
-    }
-    rewards_calculation_result.subnets_failure_rates = ctx.subnets_fr;
-
-    Ok(rewards_calculation_result)
-}
-
-fn validate_input(
-    reward_period: &RewardPeriod,
-    metrics_by_node: &BTreeMap<NodeId, Vec<NodeMetricsDaily>>,
-    all_nodes: &[NodeId],
-) -> Result<(), RewardCalculationError> {
-    if all_nodes.is_empty() {
+    input: &RewardsCalculatorInput,
+    rewardable_nodes: Vec<RewardableNode>,
+) -> Result<RewardsCalculatorResults, RewardCalculationError> {
+    if rewardable_nodes.is_empty() {
         return Err(RewardCalculationError::EmptyNodes);
     }
 
-    // Check if all nodes with metrics are present in the rewardable nodes
-    for node_id in metrics_by_node.keys() {
-        if !all_nodes.contains(node_id) {
-            return Err(RewardCalculationError::NodeNotInRewardables(*node_id));
-        }
-    }
+    // Performance Multipliers Calculation
+    let ctx: RewardsCalculator<Initialized> = RewardsCalculator {
+        input,
+        rewardable_nodes,
+        calculator_results: RewardsCalculatorResults::default(),
+        _marker: PhantomData,
+    };
+    let computed: RewardsCalculator<RewardsTotalComputed> = ctx.next().next().next().next().next().next().next().next().next().next();
+    let calculation_results = computed.get_results();
 
-    for (node_id, metrics_entries) in metrics_by_node {
-        for entry in metrics_entries {
-            // Check if all metrics are within the reward period
-            if !reward_period.contains(entry.ts.get()) {
-                return Err(RewardCalculationError::NodeMetricsOutOfRange {
-                    node_id: *node_id,
-                    timestamp: entry.ts.get(),
-                    reward_period: reward_period.clone(),
-                });
-            }
-        }
-        // Metrics are unique if there are no duplicate entries for the same day and subnet.
-        // Metrics with the same timestamp and different subnet are allowed.
-        let unique_timestamp_subnet = metrics_entries
-            .iter()
-            .map(|daily_metrics| (daily_metrics.ts.get(), daily_metrics.subnet_assigned))
-            .collect::<HashSet<_>>();
-        if unique_timestamp_subnet.len() != metrics_entries.len() {
-            return Err(RewardCalculationError::DuplicateMetrics(*node_id));
-        }
-    }
-
-    Ok(())
-}
-
-#[derive(Debug, PartialEq)]
-pub enum RewardCalculationError {
-    EmptyNodes,
-    NodeNotInRewardables(NodeId),
-    NodeMetricsOutOfRange {
-        node_id: NodeId,
-        timestamp: TimestampNanos,
-        reward_period: RewardPeriod,
-    },
-    DuplicateMetrics(NodeId),
-}
-
-impl Error for RewardCalculationError {}
-
-impl fmt::Display for RewardCalculationError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RewardCalculationError::EmptyNodes => {
-                write!(f, "No rewardable nodes provided")
-            }
-            RewardCalculationError::NodeNotInRewardables(node_id) => {
-                write!(f, "Node {} has metrics but it is not part of rewardable nodes", node_id)
-            }
-            RewardCalculationError::NodeMetricsOutOfRange {
-                node_id,
-                timestamp,
-                reward_period,
-            } => {
-                write!(
-                    f,
-                    "Node {} has metrics outside the reward period: timestamp: {} not in {}",
-                    node_id, timestamp, reward_period
-                )
-            }
-            RewardCalculationError::DuplicateMetrics(node_id) => {
-                write!(f, "Node {} has multiple metrics for the same day", node_id)
-            }
-        }
-    }
+    Ok(calculation_results)
 }
 
 #[cfg(test)]
