+++ rs/dre-canisters/node-provider-rewards/rewards-calculation/src/tests.rs
@@ -1,154 +1,38 @@
 use super::*;
-use crate::reward_period::{DayEndNanos, NANOS_PER_DAY};
-use ic_base_types::SubnetId;
-use ic_protobuf::registry::node_rewards::v2::{NodeRewardRate, NodeRewardRates};
+use crate::input_builder::{NodeMetricsDaily, RewardsCalculatorInputBuilder};
+use crate::types::{DayEndNanos, TimestampNanos, NANOS_PER_DAY};
+use ic_base_types::{NodeId, PrincipalId, SubnetId};
+use ic_protobuf::registry::node_rewards::v2::{NodeRewardRate, NodeRewardRates, NodeRewardsTable};
 use itertools::Itertools;
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
+use std::collections::BTreeMap;
 
-fn node_id(id: u64) -> NodeId {
+pub(crate) fn node_id(id: u64) -> NodeId {
     PrincipalId::new_node_test_id(id).into()
 }
 
-fn subnet_id(id: u64) -> SubnetId {
+pub(crate) fn subnet_id(id: u64) -> SubnetId {
     PrincipalId::new_subnet_test_id(id).into()
 }
 
-fn provider_id(id: u64) -> PrincipalId {
-    PrincipalId::new_user_test_id(id)
-}
-
-fn create_metrics_by_node() -> BTreeMap<NodeId, Vec<NodeMetricsDaily>> {
-    let mut metrics_by_node = BTreeMap::new();
-    metrics_by_node.insert(node_id(1), vec![NodeMetricsDaily::new(NANOS_PER_DAY, subnet_id(1), 0, 0)]);
-    metrics_by_node
-}
-
-#[test]
-fn test_empty_rewardable_nodes() {
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let result = validate_input(&reward_period, &BTreeMap::new(), &[]);
-
-    assert_eq!(result, Err(RewardCalculationError::EmptyNodes));
-}
-
-#[test]
-fn test_node_not_in_rewardables() {
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let metrics_by_node = create_metrics_by_node();
-
-    let result = validate_input(&reward_period, &metrics_by_node, &[node_id(2)]);
-    assert_eq!(result, Err(RewardCalculationError::NodeNotInRewardables(node_id(1))));
-}
-
-#[test]
-fn test_metrics_out_of_range() {
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let mut metrics_by_node = create_metrics_by_node();
-
-    let metrics_out_of_range = NodeMetricsDaily::new(0, subnet_id(1), 0, 0);
-    metrics_by_node.get_mut(&node_id(1)).unwrap().push(metrics_out_of_range.clone());
-
-    let result = validate_input(&reward_period, &metrics_by_node, &[node_id(1)]);
-
-    assert_eq!(
-        result,
-        Err(RewardCalculationError::NodeMetricsOutOfRange {
-            node_id: node_id(1),
-            timestamp: metrics_out_of_range.ts.get(),
-            reward_period,
-        })
-    );
-}
-
-#[test]
-fn test_same_day_metrics_same_sub() {
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let mut metrics_by_node = create_metrics_by_node();
-
-    metrics_by_node
-        .get_mut(&node_id(1))
-        .unwrap()
-        .push(NodeMetricsDaily::new(NANOS_PER_DAY, subnet_id(1), 0, 0));
-    let result = validate_input(&reward_period, &metrics_by_node, &[node_id(1)]);
-
-    assert_eq!(result, Err(RewardCalculationError::DuplicateMetrics(node_id(1))));
-}
-
-#[test]
-fn test_same_day_metrics_different_subs() {
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let mut metrics_by_node = create_metrics_by_node();
-
-    metrics_by_node
-        .get_mut(&node_id(1))
-        .unwrap()
-        .push(NodeMetricsDaily::new(NANOS_PER_DAY, subnet_id(2), 0, 0));
-    let result = validate_input(&reward_period, &metrics_by_node, &[node_id(1)]);
-
-    assert_eq!(result, Ok(()));
-}
-
-#[test]
-fn test_node_provider_below_min_limit() {
-    let node_provider_id = PrincipalId::new_anonymous();
-    let reward_period = RewardPeriod::new(NANOS_PER_DAY, 30 * NANOS_PER_DAY).unwrap();
-    let rewards_table = NodeRewardsTable::default();
-    let rewardables = vec![
-        RewardableNode {
-            node_id: PrincipalId::new_user_test_id(1).into(),
-            node_provider_id,
-            region: "region1".to_string(),
-            node_type: "type1".to_string(),
-        },
-        RewardableNode {
-            node_id: PrincipalId::new_user_test_id(2).into(),
-            node_provider_id,
-            region: "region1".to_string(),
-            node_type: "type3.1".to_string(),
-        },
-    ];
-
-    let rewards = calculate_rewards(&reward_period, &rewards_table, &BTreeMap::new(), &rewardables).unwrap();
-    assert_eq!(rewards.results_per_provider.get(&node_provider_id).unwrap().rewards_total, dec!(2));
-}
-
-struct NPRInput {
-    reward_period: RewardPeriod,
+#[derive(Default)]
+pub(crate) struct TestInputBuilder {
+    start_ts: TimestampNanos,
+    end_ts: TimestampNanos,
     rewards_table: NodeRewardsTable,
-    metrics_by_node: BTreeMap<NodeId, Vec<NodeMetricsDaily>>,
+    daily_metrics_by_node: BTreeMap<NodeId, Vec<NodeMetricsDaily>>,
     rewardables: Vec<RewardableNode>,
 }
 
-impl NPRInput {
-    pub fn calculate_rewards(&self) -> RewardsCalculationResult {
-        calculate_rewards(&self.reward_period, &self.rewards_table, &self.metrics_by_node, &self.rewardables).unwrap()
-    }
-}
-
-struct NPRInputBuilder {
-    reward_period: Option<RewardPeriod>,
-    rewards_table: Option<NodeRewardsTable>,
-    metrics_by_node: BTreeMap<NodeId, Vec<NodeMetricsDaily>>,
-    rewardables: Vec<RewardableNode>,
-}
-
-impl NPRInputBuilder {
-    pub fn new() -> NPRInputBuilder {
-        NPRInputBuilder {
-            reward_period: None,
-            rewards_table: None,
-            metrics_by_node: BTreeMap::new(),
-            rewardables: vec![],
-        }
-    }
-
-    pub fn with_reward_period(&mut self, start: u64, end: u64) -> &mut NPRInputBuilder {
-        self.reward_period = Some(RewardPeriod::new(start, end).unwrap());
+impl TestInputBuilder {
+    pub fn with_reward_period(mut self, start: u64, end: u64) -> Self {
+        self.start_ts = start.into();
+        self.end_ts = end.into();
         self
     }
 
-    pub fn with_rewards_rates(&mut self, region: &str, node_types: Vec<&str>, rate: u64, coeff: u64) -> &mut NPRInputBuilder {
+    pub fn with_rewards_rates(mut self, region: &str, node_types: Vec<&str>, rate: u64, coeff: u64) -> Self {
         let mut rates: BTreeMap<String, NodeRewardRate> = BTreeMap::new();
         for node_type in node_types {
             rates.insert(
@@ -159,31 +43,27 @@ impl NPRInputBuilder {
                 },
             );
         }
-        self.rewards_table
-            .get_or_insert_with(NodeRewardsTable::default)
-            .table
-            .insert(region.to_string(), NodeRewardRates { rates });
+        self.rewards_table.table.insert(region.to_string(), NodeRewardRates { rates });
         self
     }
 
-    pub fn with_node(&mut self, node_id: NodeId, node_provider_id: PrincipalId, region: &str, node_type: &str) -> &mut NPRInputBuilder {
+    pub fn with_node(mut self, node_id: NodeId, region: &str, node_type: &str) -> Self {
         self.rewardables.push(RewardableNode {
             node_id,
-            node_provider_id,
             region: region.to_string(),
             node_type: node_type.to_string(),
         });
         self
     }
 
-    pub fn with_nodes(&mut self, node_ids: Vec<NodeId>, node_provider_id: PrincipalId, region: &str, node_type: &str) -> &mut NPRInputBuilder {
+    pub fn with_nodes(mut self, node_ids: Vec<NodeId>, region: &str, node_type: &str) -> Self {
         for node_id in node_ids {
-            self.with_node(node_id, node_provider_id, region, node_type);
+            self = self.with_node(node_id, region, node_type);
         }
         self
     }
 
-    pub fn with_node_metrics(&mut self, node_id: NodeId, ts_start: u64, failure_rates: Vec<Decimal>, subnet_id: SubnetId) -> &mut NPRInputBuilder {
+    pub fn with_node_metrics(mut self, node_id: NodeId, ts_start: u64, failure_rates: Vec<Decimal>, subnet_id: SubnetId) -> Self {
         let daily_metrics: Vec<NodeMetricsDaily> = failure_rates
             .iter()
             .enumerate()
@@ -196,33 +76,30 @@ impl NPRInputBuilder {
             })
             .collect();
 
-        self.metrics_by_node
+        self.daily_metrics_by_node
             .entry(node_id)
             .and_modify(|metrics| metrics.extend(daily_metrics.iter().cloned().collect_vec()))
             .or_insert(daily_metrics);
         self
     }
 
-    pub fn with_nodes_metrics(
-        &mut self,
-        node_ids: Vec<NodeId>,
-        ts_start: u64,
-        failure_rates: Vec<Decimal>,
-        subnet_id: SubnetId,
-    ) -> &mut NPRInputBuilder {
+    pub fn with_nodes_metrics(mut self, node_ids: Vec<NodeId>, ts_start: u64, failure_rates: Vec<Decimal>, subnet_id: SubnetId) -> Self {
         for node_id in node_ids {
-            self.with_node_metrics(node_id, ts_start, failure_rates.clone(), subnet_id);
+            self = self.with_node_metrics(node_id, ts_start, failure_rates.clone(), subnet_id);
         }
         self
     }
 
-    pub fn build(&self) -> NPRInput {
-        NPRInput {
-            reward_period: self.reward_period.clone().unwrap(),
-            rewards_table: self.rewards_table.clone().unwrap(),
-            metrics_by_node: self.metrics_by_node.clone(),
-            rewardables: self.rewardables.clone(),
-        }
+    pub fn build(self) -> (RewardsCalculatorInput, Vec<RewardableNode>) {
+        (
+            RewardsCalculatorInputBuilder::default()
+                .with_start_end_ts(self.start_ts, self.end_ts)
+                .with_rewards_table(self.rewards_table)
+                .with_daily_metrics_by_node(self.daily_metrics_by_node)
+                .build()
+                .unwrap(),
+            self.rewardables,
+        )
     }
 }
 
@@ -230,25 +107,22 @@ impl NPRInputBuilder {
 fn test_node_provider_rewards_one_assigned() {
     let subnet_1 = subnet_id(1);
 
-    let np_1 = provider_id(1);
     let node_1 = node_id(1);
     let nodes_np_1 = vec![node_1, node_id(2), node_id(3), node_id(4), node_id(5)];
 
-    let np_2 = provider_id(2);
     let nodes_np_2 = vec![node_id(6), node_id(7), node_id(8)];
 
-    let input = NPRInputBuilder::new()
+    let (input, rewardables) = TestInputBuilder::default()
         .with_reward_period(0, 30 * NANOS_PER_DAY)
         .with_rewards_rates("A,B", vec!["type0", "type1", "type3"], 1000, 97)
         // Node Provider 1: node_1 assigned, rest unassigned
-        .with_nodes(nodes_np_1, np_1, "A,B", "type1")
+        .with_nodes(nodes_np_1, "A,B", "type1")
         .with_node_metrics(node_1, 0, vec![dec!(0.4), dec!(0.2), dec!(0.3), dec!(0.4)], subnet_1)
         // Node Provider 2: all assigned with 0 failure rate this for bringing the subnet failure rate to 0
-        .with_nodes(nodes_np_2.clone(), np_2, "A,B", "type1")
         .with_nodes_metrics(nodes_np_2, 0, vec![dec!(0); 4], subnet_1)
         .build();
 
-    let rewards = input.calculate_rewards();
+    let result = calculate_rewards(&input, rewardables).unwrap();
 
     //     ┌─Node: 3jo2y-lqbaa-aaaaa-aaaap-2ai ─────────────────────┬─────────────────────────────┬───────────────────────────┬─────────────────────────────┬─────────────────────────────────┐
     //     │        Day (UTC)         │ Original Failure Rate [OFR] │       Subnet Assigned       │ Subnet Failure Rate [SFR] │ Relative Failure Rate [RFR] │ Extrapolated Failure Rate [EFR] │
@@ -317,5 +191,5 @@ fn test_node_provider_rewards_one_assigned() {
     //     ├─────────────────────────────┼───────────┼─────────────┼──────────────┼────────┼        ┼────────┼────────┼────────┼────────────┼             ┤
     //     │ 2o3ay-vafaa-aaaaa-aaaap-2ai │   type1   │     A,B     │ 1000 myrXDR  │   -    │        │ 0.325  │ 0.360  │ 0.640  │ 640 myrXDR │             │
     //     └─────────────────────────────┴───────────┴─────────────┴──────────────┴────────┴────────┴────────┴────────┴────────┴────────────┴─────────────┘
-    assert_eq!(rewards.results_per_provider.get(&np_1).unwrap().rewards_total, dec!(3200));
+    assert_eq!(result.rewards_total, dec!(3200));
 }
